Extract performance tests logger
